[BEAM-545] Promote JobName to PipelineOptions
